refactor server instantiation code for supertest
